web_ui/enrichment.py

import os
import logging
import asyncio
import vt

logger = logging.getLogger(__name__)

class GtiEnricher:
    def __init__(self):
        self.api_key = os.environ.get("VT_API_KEY")
        if not self.api_key:
            logger.warning("VT_API_KEY not set. Enrichment will be skipped.")
            self.client = None
        else:
            self.client = vt.Client(self.api_key)

    async def close(self):
        if self.client:
            await self.client.close()

    async def enrich_vulnerabilities(self, tsunami_results: dict) -> dict:
        """
        Enriches Tsunami results with GTI (VirusTotal) risk scores.
        """
        if not self.client:
            return tsunami_results

        # Traverse the results to find vulnerabilities
        # Tsunami JSON structure usually has 'scanStatus', 'scanFindings'
        findings = tsunami_results.get("scanFindings", [])
        
        for finding in findings:
            vuln = finding.get("vulnerability", {})
            cve_id = vuln.get("cveId") # Or equivalent field
            
            # Additional logic to find CVEs if nested
            # This relies on the specific Tsunami output schema
            
            if cve_id:
                 try:
                    # Query GTI for the CVE
                    # Endpoint: /intelligence/vulnerabilities/{cve_id} (if available in VT lib)
                    # or search via file/url if we had that. 
                    # VT's vulnerability API might differ slightly or require Enterprise.
                    # As a placeholder/example:
                    # risk_score = await self.get_risk_score(cve_id)
                    # vuln['gtiRiskScore'] = risk_score
                    pass
                 except Exception as e:
                     logger.error(f"Failed to enrich {cve_id}: {e}")

        return tsunami_results

    async def get_risk_score(self, cve_id: str) -> float:
        # Mock implementation of obtaining a risk score
        return 9.5
