web_ui/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
google-cloud-run==0.10.0
google-cloud-storage==2.14.0
python-multipart==0.0.7
vt-py==0.18.0

